Fixed removal of epsilon productions
And this is why I love this automatic stuff.

Turns out I was only removing *one* nullable variable per production (at
a time) so for instance

A -> BC
would become
A -> B and
A -> C
but not
A -> e

This was solved using the standard method of copying data then seeing if
it changes.